Remove GUI dependency from helpdata.h
The help generation code is used by civmanual, which has to link to libclient
(and provide shim implementation of some client functions) because helpdata
calls into the client code directly.

Start reducing the client dependencies in helpdata by making its header not
depend on client code.

See #915.
