+++ client/helpdata.h
@@ -12,7 +12,56 @@
       \____/        ********************************************************/
 #pragma once
 
-#include "helpdlg_g.h" // enum help_page_type
+// This must be in same order as names in helpdata.cpp
+enum help_page_type {
+  HELP_ANY,
+  HELP_TEXT,
+  HELP_UNIT,
+  HELP_IMPROVEMENT,
+  HELP_WONDER,
+  HELP_TECH,
+  HELP_TERRAIN,
+  HELP_EXTRA,
+  HELP_GOODS,
+  HELP_SPECIALIST,
+  HELP_GOVERNMENT,
+  HELP_RULESET,
+  HELP_TILESET,
+  HELP_NATIONS,
+  HELP_MULTIPLIER,
+  HELP_EFFECT,
+  HELP_LAST
+};
+
+// TRANS: "Overview" topic in built-in help
+#define HELP_OVERVIEW_ITEM N_("?help:Overview")
+#define HELP_PLAYING_ITEM N_("Strategy and Tactics")
+#define HELP_LANGUAGES_ITEM N_("Languages")
+#define HELP_CONNECTING_ITEM N_("Connecting")
+#define HELP_CHATLINE_ITEM N_("Chatline")
+#define HELP_WORKLIST_EDITOR_ITEM N_("Worklist Editor")
+#define HELP_CMA_ITEM N_("Citizen Governor")
+#define HELP_CONTROLS_ITEM N_("Controls")
+#define HELP_RULESET_ITEM N_("About Current Ruleset")
+#define HELP_TILESET_ITEM N_("About Current Tileset")
+#define HELP_NATIONS_ITEM N_("About Nations")
+#define HELP_ECONOMY_ITEM N_("Economy")
+#define HELP_CITIES_ITEM N_("Cities")
+#define HELP_IMPROVEMENTS_ITEM N_("City Improvements")
+#define HELP_UNITS_ITEM N_("Units")
+#define HELP_COMBAT_ITEM N_("Combat")
+#define HELP_ZOC_ITEM N_("Zones of Control")
+#define HELP_TECHS_ITEM N_("Technology")
+#define HELP_EXTRAS_ITEM N_("Extras")
+#define HELP_TERRAIN_ITEM N_("Terrain")
+#define HELP_WONDERS_ITEM N_("Wonders of the World")
+#define HELP_GOVERNMENT_ITEM N_("Government")
+#define HELP_EFFECTS_ITEM N_("Effects")
+#define HELP_DIPLOMACY_ITEM N_("Diplomacy")
+#define HELP_SPACE_RACE_ITEM N_("Space Race")
+#define HELP_COPYING_ITEM N_("Copying")
+#define HELP_ABOUT_ITEM N_("About Freeciv21")
+#define HELP_MULTIPLIER_ITEM N_("Policies")
 
 struct help_item {
   char *topic, *text;
